requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.3.0
setuptools>=65.0.0
